mito-ai: cleanup
